+++ Sources/StreamChatSwiftUI/ChatChannel/MessageList/LinkAttachmentView.swift
@@ -49,13 +49,22 @@ public struct LinkAttachmentContainer<Factory: ViewFactory>: View {
 
             let availableWidth = width - 4 * padding
             let size = message.adjustedText.frameSize(maxWidth: availableWidth)
-            LinkTextView(
-                message: message,
-                width: availableWidth,
-                textColor: UIColor(textColor(for: message))
-            )
-            .frame(width: availableWidth, height: size.height)
-            .standardPadding()
+            
+            if #available(iOS 15, *) {
+                HStack {
+                    StreamTextView(message: message)
+                        .standardPadding()
+                    Spacer()
+                }
+            } else {
+                LinkTextView(
+                    message: message,
+                    width: availableWidth,
+                    textColor: UIColor(textColor(for: message))
+                )
+                .frame(width: availableWidth, height: size.height)
+                .standardPadding()
+            }
 
             if !message.linkAttachments.isEmpty {
                 LinkAttachmentView(