+++ StreamChatSwiftUITests/Tests/ChatChannel/MessageView_Tests.swift
@@ -387,4 +387,216 @@ class MessageView_Tests: StreamChatTestCase {
         // When
         assertSnapshot(matching: view, as: .image(perceptualPrecision: precision))
     }
+    
+    func test_markdown_noLinks() {
+        // Given
+        let textMessage = ChatMessage.mock(
+            id: .unique,
+            cid: .unique,
+            text: "This is a **bold** text",
+            author: .mock(id: .unique)
+        )
+
+        // When
+        let view = MessageView(
+            factory: DefaultViewFactory.shared,
+            message: textMessage,
+            contentWidth: defaultScreenSize.width,
+            isFirst: true,
+            scrolledId: .constant(nil)
+        )
+        .frame(
+            width: defaultScreenSize.width,
+            height: 100
+        )
+
+        // Then
+        assertSnapshot(matching: view, as: .image(perceptualPrecision: precision))
+    }
+    
+    func test_markdown_withLinks() {
+        // Given
+        let textMessage = ChatMessage.mock(
+            id: .unique,
+            cid: .unique,
+            text: "Visit Apple, click [here](https://apple.com)",
+            author: .mock(id: .unique)
+        )
+
+        // When
+        let view = MessageView(
+            factory: DefaultViewFactory.shared,
+            message: textMessage,
+            contentWidth: defaultScreenSize.width,
+            isFirst: true,
+            scrolledId: .constant(nil)
+        )
+        .frame(
+            width: defaultScreenSize.width,
+            height: 100
+        )
+
+        // Then
+        assertSnapshot(matching: view, as: .image(perceptualPrecision: precision))
+    }
+    
+    func test_markdown_withLinksDifferentTint() {
+        // Given
+        let displayOptions = MessageDisplayOptions { _ in
+            [
+                NSAttributedString.Key.foregroundColor: UIColor.red
+            ]
+        }
+        let config = MessageListConfig(messageDisplayOptions: displayOptions)
+        let utils = Utils(messageListConfig: config)
+        streamChat = StreamChat(chatClient: chatClient, utils: utils)
+        
+        let textMessage = ChatMessage.mock(
+            id: .unique,
+            cid: .unique,
+            text: "Visit Apple, click [here](https://apple.com)",
+            author: .mock(id: .unique)
+        )
+
+        // When
+        let view = MessageView(
+            factory: DefaultViewFactory.shared,
+            message: textMessage,
+            contentWidth: defaultScreenSize.width,
+            isFirst: true,
+            scrolledId: .constant(nil)
+        )
+        .frame(
+            width: defaultScreenSize.width,
+            height: 100
+        )
+
+        // Then
+        assertSnapshot(matching: view, as: .image(perceptualPrecision: precision))
+    }
+    
+    func test_markdown_disabledWithLinks() {
+        // Given
+        let config = MessageListConfig(markdownSupportEnabled: false)
+        let utils = Utils(messageListConfig: config)
+        streamChat = StreamChat(chatClient: chatClient, utils: utils)
+        
+        let textMessage = ChatMessage.mock(
+            id: .unique,
+            cid: .unique,
+            text: "Visit Apple, click [here](https://apple.com)",
+            author: .mock(id: .unique)
+        )
+
+        // When
+        let view = MessageView(
+            factory: DefaultViewFactory.shared,
+            message: textMessage,
+            contentWidth: defaultScreenSize.width,
+            isFirst: true,
+            scrolledId: .constant(nil)
+        )
+        .frame(
+            width: defaultScreenSize.width,
+            height: 100
+        )
+
+        // Then
+        assertSnapshot(matching: view, as: .image(perceptualPrecision: precision))
+    }
+    
+    func test_markdown_disabledWithRegularText() {
+        // Given
+        let config = MessageListConfig(markdownSupportEnabled: false)
+        let utils = Utils(messageListConfig: config)
+        streamChat = StreamChat(chatClient: chatClient, utils: utils)
+        
+        let textMessage = ChatMessage.mock(
+            id: .unique,
+            cid: .unique,
+            text: "This is a **bold** text",
+            author: .mock(id: .unique)
+        )
+
+        // When
+        let view = MessageView(
+            factory: DefaultViewFactory.shared,
+            message: textMessage,
+            contentWidth: defaultScreenSize.width,
+            isFirst: true,
+            scrolledId: .constant(nil)
+        )
+        .frame(
+            width: defaultScreenSize.width,
+            height: 100
+        )
+
+        // Then
+        assertSnapshot(matching: view, as: .image(perceptualPrecision: precision))
+    }
+    
+    func test_markdownAndLinkSupport_disabled() {
+        // Given
+        let config = MessageListConfig(
+            localLinkDetectionEnabled: false,
+            markdownSupportEnabled: false
+        )
+        let utils = Utils(messageListConfig: config)
+        streamChat = StreamChat(chatClient: chatClient, utils: utils)
+        
+        let textMessage = ChatMessage.mock(
+            id: .unique,
+            cid: .unique,
+            text: "Visit Apple, click [here](https://apple.com)",
+            author: .mock(id: .unique)
+        )
+
+        // When
+        let view = MessageView(
+            factory: DefaultViewFactory.shared,
+            message: textMessage,
+            contentWidth: defaultScreenSize.width,
+            isFirst: true,
+            scrolledId: .constant(nil)
+        )
+        .frame(
+            width: defaultScreenSize.width,
+            height: 100
+        )
+
+        // Then
+        assertSnapshot(matching: view, as: .image(perceptualPrecision: precision))
+    }
+    
+    func test_markdown_linkSupportDisabled() {
+        // Given
+        let config = MessageListConfig(
+            localLinkDetectionEnabled: false
+        )
+        let utils = Utils(messageListConfig: config)
+        streamChat = StreamChat(chatClient: chatClient, utils: utils)
+        
+        let textMessage = ChatMessage.mock(
+            id: .unique,
+            cid: .unique,
+            text: "~~A strikethrough example~~",
+            author: .mock(id: .unique)
+        )
+
+        // When
+        let view = MessageView(
+            factory: DefaultViewFactory.shared,
+            message: textMessage,
+            contentWidth: defaultScreenSize.width,
+            isFirst: true,
+            scrolledId: .constant(nil)
+        )
+        .frame(
+            width: defaultScreenSize.width,
+            height: 100
+        )
+
+        // Then
+        assertSnapshot(matching: view, as: .image(perceptualPrecision: precision))
+    }
 }